Move FixtureProvider class and tests to own dir
