+++ packages/react-cosmos-render/src/FixtureProvider/__tests__/_shared.js
@@ -1,9 +1,9 @@
 // @flow
 
-import { extractValuesFromObject } from '../shared/values';
-import { setFixtureUpdater } from '../shared/fixture-state';
+import { extractValuesFromObject } from '../../shared/values';
+import { setFixtureUpdater } from '../../shared/fixture-state';
 
-import type { FixtureState, FixtureStateUpdater } from '../types';
+import type { FixtureState, FixtureStateUpdater } from '../../types';
 
 export function updateFixtureState(
   prevState: FixtureState,
